Edit button launches modal
