+++ api/database.json
@@ -99,8 +99,8 @@
       "title": "BOAT CRASH",
       "summary": "TITANIC 2!?!?!?",
       "url": "https://www.testitout.com",
-      "userId": 1,
-      "timestamp": "2018-11-14T23:02:52.195Z",
+      "userId": 2,
+      "timestamp": "2018-11-29T16:54:40.372Z",
       "id": 1
     },
     {
