+++ src/components/news/News.js
@@ -3,6 +3,7 @@ import { Container } from 'reactstrap'
 import NewsList from './NewsList'
 import NewsModal from './NewsModal'
 import API from './../../modules/API/API'
+import moment from 'moment'
 
 export default class News extends React.Component {
 
@@ -18,12 +19,38 @@ export default class News extends React.Component {
       .then(() => this.setState(newState))
   }
 
+  // Posts new article to database and adds them to state.
   saveArticle = (articleInfo) => {
     API.saveData("news", articleInfo)
       .then(() => {
       // Clears main container and pulls new news dashboard with additional article.
       return this.newsLog()
     })
+  }
+
+  editArticle = (articleInfo, id) => {
+    API.editData("news", articleInfo, id)
+      .then(() => {
+      // Clears main container and pulls new news dashboard with additional article.
+      return this.newsLog()
+    })
+  }
+
+  // OnClick functionality to edit saved articles
+  handleEdit = (id) => {
+    // ADD TIMESTAMP
+    let timeSaved = moment(new Date())
+    // ADD USER ID ON CLICK - Currently hard coded, needs to be userSession.getUser()
+    let userId = 2
+    let articleInfo = {
+      title: this.state.title,
+      summary: this.state.summary,
+      url: this.state.url,
+      userId: userId,
+      timestamp: timeSaved
+    }
+    console.log(articleInfo)
+    this.editArticle(articleInfo, id)
   }
 
   // OnClick functionality to delete saved articles
@@ -43,7 +70,7 @@ export default class News extends React.Component {
       <Container>
         <h1 className="text-center mt-5">News Around the 'Berg!</h1>
         <NewsModal saveArticle={this.saveArticle} />
-        <NewsList news={this.state.news} handleDelete={this.handleDelete} />
+        <NewsList news={this.state.news} handleDelete={this.handleDelete} handleEdit={this.handleEdit}/>
       </Container>)
   }
 }
